Add PubNub integration
Add PubNub integration to send a message to PubNub everytime
the doorbell button is pressed. This version will send the
timestamp of the time the button is pressed with the name of the
person that pressed the button. If the person is not recognized,
it will send 'Unknown'.

This can be used with the mobile app to receive real time
notification with each doorbell press.